+++ HARK/ConsumptionSaving/ConsPortfolioModel.py
@@ -583,9 +583,7 @@ class ConsPortfolioSolver(MetricObject):
         dvdb_intermed = calc_expectation(
             self.IncShkDstn, dvdb_dist, self.bNrmNext, self.ShareNext
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        dvdb_intermed = dvdb_intermed[:, :, 0]
+        
         dvdbNvrs_intermed = self.uPinv(dvdb_intermed)
         dvdbNvrsFunc_intermed = BilinearInterp(
             dvdbNvrs_intermed, self.bNrmGrid, self.ShareGrid
@@ -596,9 +594,7 @@ class ConsPortfolioSolver(MetricObject):
         dvds_intermed = calc_expectation(
             self.IncShkDstn, dvds_dist, self.bNrmNext, self.ShareNext
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        dvds_intermed = dvds_intermed[:, :, 0]
+        
         dvdsFunc_intermed = BilinearInterp(dvds_intermed, self.bNrmGrid, self.ShareGrid)
 
         # Make tiled arrays to calculate future realizations of bNrm and Share when integrating over RiskyDstn
@@ -636,9 +632,7 @@ class ConsPortfolioSolver(MetricObject):
                 self.RiskyDstn, EndOfPrddvda_dist, self.aNrm_tiled, self.ShareNext
             )
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        self.EndOfPrddvda = self.EndOfPrddvda[:, :, 0]
+
         self.EndOfPrddvdaNvrs = self.uPinv(self.EndOfPrddvda)
 
         # Calculate end-of-period marginal value of risky portfolio share by taking expectations
@@ -649,9 +643,6 @@ class ConsPortfolioSolver(MetricObject):
                 self.RiskyDstn, EndOfPrddvds_dist, self.aNrm_tiled, self.ShareNext
             )
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        self.EndOfPrddvds = self.EndOfPrddvds[:, :, 0]
 
     def optimize_share(self):
         """
@@ -802,9 +793,7 @@ class ConsPortfolioSolver(MetricObject):
         v_intermed = calc_expectation(
             self.IncShkDstn, v_intermed_dist, self.bNrmNext, self.ShareNext
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        v_intermed = v_intermed[:, :, 0]
+
         vNvrs_intermed = self.uinv(v_intermed)
         vNvrsFunc_intermed = BilinearInterp(
             vNvrs_intermed, self.bNrmGrid, self.ShareGrid
@@ -828,9 +817,7 @@ class ConsPortfolioSolver(MetricObject):
                 self.RiskyDstn, EndOfPrdv_dist, self.aNrm_tiled, self.ShareNext
             )
         )
-        # calc_expectation returns one additional "empty" dimension, remove it
-        # this line can be deleted when calc_expectation is fixed
-        self.EndOfPrdv = self.EndOfPrdv[:, :, 0]
+
         self.EndOfPrdvNvrs = self.uinv(self.EndOfPrdv)
 
     def make_vFunc(self):
@@ -1143,7 +1130,7 @@ class ConsPortfolioJointDistSolver(ConsPortfolioDiscreteSolver, ConsPortfolioSol
                 self.ShockDstn, EndOfPrddvda_dists, self.aNrm_tiled, self.Share_tiled
             )
         )
-        self.EndOfPrddvda = self.EndOfPrddvda[:, :, 0]
+
         self.EndOfPrddvdaNvrs = self.uPinv(self.EndOfPrddvda)
 
         # Calculate end-of-period marginal value of risky portfolio share by taking expectations
@@ -1154,7 +1141,6 @@ class ConsPortfolioJointDistSolver(ConsPortfolioDiscreteSolver, ConsPortfolioSol
                 self.ShockDstn, EndOfPrddvds_dist, self.aNrm_tiled, self.Share_tiled
             )
         )
-        self.EndOfPrddvds = self.EndOfPrddvds[:, :, 0]
 
     def make_EndOfPrdvFunc(self):
         """
@@ -1182,7 +1168,7 @@ class ConsPortfolioJointDistSolver(ConsPortfolioDiscreteSolver, ConsPortfolioSol
                 self.ShockDstn, v_dist, self.aNrm_tiled, self.Share_tiled
             )
         )
-        self.EndOfPrdv = self.EndOfPrdv[:, :, 0]
+
         self.EndOfPrdvNvrs = self.uinv(self.EndOfPrdv)
 
     def solve(self):